Convert session_controller functional tests
